Move interaction functions to the respective modules
