+++ src/pykeen/models/base.py
@@ -1204,89 +1204,3 @@ class InteractionFunction(nn.Module):
                 continue
             if hasattr(mod, 'reset_parameters'):
                 mod.reset_parameters()
-
-
-def _normalize_for_einsum(
-    x: torch.FloatTensor,
-    batch_size: int,
-    symbol: str,
-) -> Tuple[str, torch.FloatTensor]:
-    """
-    Normalize tensor for broadcasting along batch-dimension in einsum.
-
-    :param x:
-        The tensor.
-    :param batch_size:
-        The batch_size
-    :param symbol:
-        The symbol for the einsum term.
-
-    :return:
-        A tuple (reshaped_tensor, term).
-    """
-    if x.shape[0] == batch_size:
-        return f'b{symbol}d', x
-    return f'{symbol}d', x.squeeze(dim=0)
-
-
-class DistMultInteractionFunction(InteractionFunction):
-    """Interaction function of DistMult."""
-
-    def forward(
-        self,
-        h: torch.FloatTensor,
-        r: torch.FloatTensor,
-        t: torch.FloatTensor,
-    ) -> torch.FloatTensor:  # noqa: D102
-        batch_size = max(h.shape[0], r.shape[0], t.shape[0])
-        h_term, h = _normalize_for_einsum(x=h, batch_size=batch_size, symbol='h')
-        r_term, r = _normalize_for_einsum(x=r, batch_size=batch_size, symbol='r')
-        t_term, t = _normalize_for_einsum(x=t, batch_size=batch_size, symbol='t')
-        return torch.einsum(f'{h_term},{r_term},{t_term}->bhrt', h, r, t)
-
-
-class ERMLPInteractionFunction(InteractionFunction):
-    """
-    Interaction function of ER-MLP.
-
-    .. math ::
-        f(h, r, t) = W_2 ReLU(W_1 cat(h, r, t) + b_1) + b_2
-    """
-
-    def __init__(
-        self,
-        embedding_dim: int,
-        hidden_dim: int,
-    ):
-        """
-        Initialize the interaction function.
-
-        :param embedding_dim:
-            The embedding vector dimension.
-        :param hidden_dim:
-            The hidden dimension of the MLP.
-        """
-        super().__init__()
-        """The multi-layer perceptron consisting of an input layer with 3 * self.embedding_dim neurons, a  hidden layer
-           with self.embedding_dim neurons and output layer with one neuron.
-           The input is represented by the concatenation embeddings of the heads, relations and tail embeddings.
-        """
-        self.head_to_hidden = nn.Linear(in_features=embedding_dim, out_features=hidden_dim, bias=False)
-        self.rel_to_hidden = nn.Linear(in_features=embedding_dim, out_features=hidden_dim, bias=True)
-        self.tail_to_hidden = nn.Linear(in_features=embedding_dim, out_features=hidden_dim, bias=False)
-        self.activation = nn.ReLU()
-        self.hidden_to_score = nn.Linear(in_features=hidden_dim, out_features=1, bias=True)
-
-    def forward(
-        self,
-        h: torch.FloatTensor,
-        r: torch.FloatTensor,
-        t: torch.FloatTensor,
-    ) -> torch.FloatTensor:  # noqa: D102
-        h = self.head_to_hidden(h)
-        r = self.rel_to_hidden(r)
-        t = self.tail_to_hidden(t)
-        # TODO: Choosing which to combine first, h/r, h/t or r/t, depending on the shape might further improve
-        #       performance in a 1:n scenario.
-        x = self.activation(h[:, :, None, None, :] + r[:, None, :, None, :] + t[:, None, None, :, :])
-        return self.hidden_to_score(x).squeeze(dim=-1)
